Added interaction with stock database for home screen
This isn't the end version of how the home screen will display information, just wanted to show that I am able to get the data from the database and display it to the user. So currently whenever you load the homescreen it will display all of the keys that are located under calls. This is dynamic, so it will update as the database updates.